requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
motor==3.3.1
pymongo==4.6.1
python-dateutil==2.8.2
aiohttp==3.9.1
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
pytz==2023.3
tenacity==8.2.3
ujson==5.9.0
